+++ dev/poc/prepare_time_combined_dataset.py
@@ -118,10 +118,12 @@ def main(cmdline=1, **kwargs):
                 --remove_seasons=$remove_seasons_str \
                 --merge_method=$merge_method \
                 --spatial_tile_size=$spatial_tile_size \
-                --start_time=2010-01-01 \
+                --start_time=2010-03-01 \
                 --assets_dname="raw_bands" \
                 --workers=$WORKERS
             '''), fmtdict)
+        if 1:
+            code = subtemplate('test -e "$OUTPUT_BUNDLE_DPATH/imgonly-${REGION}.kwcoco.zip" || ', fmtdict) + code
         combine_job = queue.submit(code, name=f'combine-time-{region}')
 
         if config.reproject:
@@ -137,6 +139,7 @@ def main(cmdline=1, **kwargs):
                 python -m watch reproject \
                     --src $OUTPUT_BUNDLE_DPATH/imgonly-${REGION}.kwcoco.zip \
                     --dst $OUTPUT_BUNDLE_DPATH/imganns-${REGION}.kwcoco.zip \
+                    --status_to_catname="positive_excluded: positive" \
                     --site_models="$TRUE_SITE_DPATH/${REGION}_*.geojson"
                 ''', fmtdict)
             queue.submit(code, depends=[field_job], name=f'reproject-ann-{region}')
@@ -157,7 +160,7 @@ if __name__ == '__main__':
                     KR_R001,
                     KR_R002, LT_R001, NZ_R001, US_R001, US_R004, US_R005,
                     US_R006, US_R007,
-                    # iMerit Regions
+                    # # iMerit Regions
                     AE_C001,
                     AE_C002,
                     AE_C003, PE_C001, QA_C001, SA_C005, US_C000, US_C010,
@@ -167,15 +170,71 @@ if __name__ == '__main__':
             --output_bundle_dpath=$DVC_DATA_DPATH/Drop6-MedianSummer10GSD \
             --true_site_dpath=$DVC_DATA_DPATH/annotations/drop6_hard_v1/site_models \
             --true_region_dpath=$DVC_DATA_DPATH/annotations/drop6_hard_v1/region_models \
-            --backend=tmux \
             --spatial_tile_size=256 \
             --merge_method=median \
             --remove_seasons=spring,fall,winter \
-            --tmux_workers=4 \
-            --time_window=3m \
-            --combine_workers=2 \
+            --tmux_workers=2 \
+            --time_window=1y \
+            --combine_workers=4 \
             --resolution=10GSD \
+            --backend=tmux \
             --run=1
+
+        # Drop 6
+        export CUDA_VISIBLE_DEVICES="0,1"
+        DVC_DATA_DPATH=$(geowatch_dvc --tags='phase2_data' --hardware=auto)
+        DVC_EXPT_DPATH=$(geowatch_dvc --tags='phase2_expt' --hardware='auto')
+        BUNDLE_DPATH=$DVC_DATA_DPATH/Drop6-MedianSummer10GSD
+        python -m watch.cli.prepare_teamfeats \
+            --base_fpath "$BUNDLE_DPATH"/imganns-*[0-9].kwcoco.zip \
+            --expt_dvc_dpath="$DVC_EXPT_DPATH" \
+            --with_landcover=1 \
+            --with_invariants2=1 \
+            --with_materials=0 \
+            --with_depth=0 \
+            --with_cold=0 \
+            --skip_existing=1 \
+            --assets_dname=teamfeats \
+            --gres=0, --tmux_workers=1 --backend=tmux --run=1
+
+        DVC_DATA_DPATH=$(geowatch_dvc --tags='phase2_data' --hardware=auto)
+        python -m watch.cli.prepare_splits \
+            --base_fpath=$DVC_DATA_DPATH/Drop6-MedianSummer10GSD/combo_imganns*_L*.kwcoco.zip \
+            --constructive_mode=True \
+            --suffix=L \
+            --backend=tmux --workers=6 \
+            --run=1
+
+        DVC_DATA_DPATH=$(geowatch_dvc --tags='phase2_data' --hardware=auto)
+        python -m watch.cli.prepare_splits \
+            --base_fpath=$DVC_DATA_DPATH/Drop6-MedianSummer10GSD/combo_imganns*_I2L*.kwcoco.zip \
+            --constructive_mode=True \
+            --suffix=I2L \
+            --backend=tmux --tmux_workers=6 \
+            --run=1
+
+        DVC_DATA_DPATH=$(geowatch_dvc --tags='phase2_data' --hardware=auto)
+        python -m watch.cli.prepare_splits \
+            --base_fpath=$DVC_DATA_DPATH/Drop6-MedianSummer10GSD/imganns-*.kwcoco.zip \
+            --constructive_mode=True \
+            --suffix=rawbands \
+            --backend=tmux --tmux_workers=6 \
+            --run=1
+
+        DVC_DATA_DPATH=$(geowatch_dvc --tags='phase2_data' --hardware=auto)
+        DVC_EXPT_DPATH=$(geowatch_dvc --tags='phase2_expt' --hardware='auto')
+        TRUE_SITE_DPATH=$DVC_DATA_DPATH/annotations/drop6_hard_v1/site_models
+        OUTPUT_BUNDLE_DPATH=$DVC_DATA_DPATH/Drop6-MeanYear10GSD-V2
+        geowatch reproject \
+            --src $DVC_DATA_DPATH/Drop6-MedianSummer10GSD/data_vali_L_split6.kwcoco.zip \
+            --dst $DVC_DATA_DPATH/Drop6-MedianSummer10GSD/data_vali_L_split6.kwcoco.zip \
+            --status_to_catname="positive_excluded: positive" \
+            --site_models=$TRUE_SITE_DPATH
+        geowatch reproject \
+            --src $DVC_DATA_DPATH/Drop6-MedianSummer10GSD/data_train_L_split6.kwcoco.zip \
+            --dst $DVC_DATA_DPATH/Drop6-MedianSummer10GSD/data_train_L_split6.kwcoco.zip \
+            --status_to_catname="positive_excluded: positive" \
+            --site_models=$TRUE_SITE_DPATH
 
         DVC_DATA_DPATH=$(smartwatch_dvc --tags='phase2_data' --hardware=auto)
         python ~/code/watch/dev/poc/prepare_time_combined_dataset.py \