+++ watch/cli/coco_add_watch_fields.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 """
 Adds fields needed by ndsampler to correctly "watch" a region.
 
@@ -5,6 +6,9 @@ Some of this is done hueristically. We assume images come from certain sensors.
 We assume input is orthorectified.  We assume some GSD "target" gsd for video
 and image processing. Note a video GSD will typically be much higher (i.e.
 lower resolution) than an image GSD.
+
+SeeAlso:
+    ~/code/watch/watch/utils/kwcoco_extensions.py
 """
 import ubelt as ub
 import scriptconfig as scfg
@@ -15,36 +19,34 @@ class AddWatchFieldsConfig(scfg.DataConfig):
     Updates image transforms in a kwcoco json file to align all videos to a
     target GSD.
     """
-    __default__ = {
-        'src': scfg.Value('data.kwcoco.json', help='input kwcoco filepath', position=1),
+    src = scfg.Value('data.kwcoco.json', help='input kwcoco filepath', position=1)
 
-        'dst': scfg.Value(None, help='output kwcoco filepath', position=2),
+    dst = scfg.Value(None, help='output kwcoco filepath', position=2)
 
-        'inplace': scfg.Value(False, isflag=True, help=ub.paragraph(
-            '''
-            if True and dst is unspecified then the output will overwrite the input
-            ''')),
+    inplace = scfg.Value(False, isflag=True, help=ub.paragraph(
+        '''
+        if True and dst is unspecified then the output will overwrite the input
+        '''))
 
-        'target_gsd': scfg.Value(10.0, help='compute transforms for a target gsd'),
+    target_gsd = scfg.Value(10.0, help='compute transforms for a target gsd')
 
-        'overwrite': scfg.Value(False, help='if True overwrites introspectable fields'),
+    overwrite = scfg.Value(False, help='if True overwrites introspectable fields')
 
-        'edit_geotiff_metadata': scfg.Value(False, help='if True MODIFIES THE UNDERLYING IMAGES to ensure geodata is propogated'),
+    edit_geotiff_metadata = scfg.Value(False, help='if True MODIFIES THE UNDERLYING IMAGES to ensure geodata is propogated')
 
-        'default_gsd': scfg.Value(None, help='if specified, assumed any images without geo-metadata have this GSD'),
+    default_gsd = scfg.Value(None, help='if specified, assumed any images without geo-metadata have this GSD')
 
-        'workers': scfg.Value(0, type=str, help='number of io threads'),
+    workers = scfg.Value(0, type=str, help='number of io threads')
 
-        'mode': scfg.Value('process', help='can be thread, process, or serial'),
+    mode = scfg.Value('process', help='can be thread, process, or serial')
 
-        'enable_video_stats': scfg.Value(True, help='set to False to disable video stats'),
+    enable_video_stats = scfg.Value(True, help='set to False to disable video stats')
 
-        'enable_valid_region': scfg.Value(False, help='set to True to enable valid region computation'),
+    enable_valid_region = scfg.Value(False, help='set to True to enable valid region computation')
 
-        'enable_intensity_stats': scfg.Value(False, help='if True, will compute intensity statistics on each channel of each image'),
+    enable_intensity_stats = scfg.Value(False, help='if True, will compute intensity statistics on each channel of each image')
 
-        'remove_broken': scfg.Value(False, help='if True, will remove any image that fails population (e.g. caused by a 404)')
-    }
+    remove_broken = scfg.Value(False, help='if True, will remove any image that fails population (e.g. caused by a 404)')
 
 
 def main(cmdline=True, **kwargs):
@@ -74,8 +76,8 @@ def main(cmdline=True, **kwargs):
             --dst=$HOME/data/dvc-repos/smart_watch_dvc/drop0_aligned_msi/data.kwcoco.new.json \
             --target_gsd=10
 
-    jq .images[0].auxiliary[0] $HOME/data/dvc-repos/smart_watch_dvc/drop0_aligned_msi/data.kwcoco.new.json
-    jq .images[0].auxiliary[0] $HOME/data/dvc-repos/smart_watch_dvc/drop0_aligned_msi/data.kwcoco.json
+        jq .images[0].auxiliary[0] $HOME/data/dvc-repos/smart_watch_dvc/drop0_aligned_msi/data.kwcoco.new.json
+        jq .images[0].auxiliary[0] $HOME/data/dvc-repos/smart_watch_dvc/drop0_aligned_msi/data.kwcoco.json
 
     Example:
         >>> from watch.cli.coco_add_watch_fields import *  # NOQA
@@ -84,27 +86,19 @@ def main(cmdline=True, **kwargs):
         >>> dset = kwcoco.CocoDataset.demo('vidshapes8-multispectral')
         >>> print('dset = {!r}'.format(dset))
         >>> target_gsd = 13.0
-        >>> main(cmdline=False, src=dset, target_gsd=target_gsd, default_gsd=1)
+        >>> main(cmdline=False, src=dset, dst='return', target_gsd=target_gsd, default_gsd=1)
         >>> print('dset.index.imgs[1] = ' + ub.urepr(dset.index.imgs[1], nl=2))
         >>> print('dset.index.videos = {}'.format(ub.urepr(dset.index.videos, nl=1)))
-
-    Ignore:
-        kwargs = {
-            'src': ub.expandpath('~/data/dvc-repos/smart_watch_dvc/drop0_aligned/data.kwcoco.json'),
-            'target_gsd': 10.0,
-            'dst': None,
-        }
-        kwargs['src'] = kwargs['dst']
-        main(**kwargs)
     """
+    import rich
+    config = AddWatchFieldsConfig.cli(default=kwargs, cmdline=cmdline, strict=True)
+    rich.print('config = {}'.format(ub.urepr(config, nl=1)))
     from watch import heuristics
     import numpy as np
     import kwcoco
     import kwimage
     from watch.utils import util_parallel
     from watch.utils import kwcoco_extensions
-    config = AddWatchFieldsConfig(kwargs, cmdline=cmdline)
-    print('config = {}'.format(ub.urepr(dict(config), nl=1)))
 
     if config['dst'] is None:
         if config['inplace']:
@@ -145,7 +139,9 @@ def main(cmdline=True, **kwargs):
                 print('offset = {!r}'.format(offset))
                 print('{}, {}'.format(gid, img['warp_img_to_vid']))
 
-    if config['dst'] is not None:
+    if config['dst'] == 'return':
+        return dset
+    elif config['dst'] is not None:
         print('write dataset')
         dset.fpath = config['dst']
         print('dset.fpath = {!r}'.format(dset.fpath))