+++ watch/stac/stac_search_builder.py
@@ -19,7 +19,7 @@ import ubelt as ub
 import scriptconfig as scfg
 
 
-class StacSearchBuilderConfig(scfg.Config):
+class StacSearchBuilderConfig(scfg.DataConfig):
     """
     Helper to create STAC search json queries
     """
@@ -252,6 +252,16 @@ _PUBLIC_L2_PRODUCTS = {
 # https://planetarycomputer.microsoft.com/dataset/landsat-c2-l2
 _PUBLIC_L2_PRODUCTS['planetarycomputer'] = {
     'collections': ['landsat-c2-l2', 'sentinel-2-l2a'],
+    "endpoint": "https://planetarycomputer.microsoft.com/api/stac/v1",
+}
+
+_PUBLIC_L2_PRODUCTS['planetarycomputer_l8'] = {
+    'collections': ['landsat-c2-l2'],
+    "endpoint": "https://planetarycomputer.microsoft.com/api/stac/v1",
+}
+
+_PUBLIC_L2_PRODUCTS['planetarycomputer_s2'] = {
+    'collections': ['sentinel-2-l2a'],
     "endpoint": "https://planetarycomputer.microsoft.com/api/stac/v1",
 }
 
@@ -495,7 +505,8 @@ def main(cmdline=1, **kwargs):
         >>> main(cmdline=cmdline, **kwargs)
     """
     import json
-    config = StacSearchBuilderConfig(cmdline=cmdline, data=kwargs)
+    config = StacSearchBuilderConfig.cli(cmdline=cmdline, data=kwargs,
+                                         strict=True)
 
     search_json = build_search_json(**ub.compatible(config, build_search_json))
     text = json.dumps(search_json, indent='    ')